webapp/requirements.txt
flask==3.1.2
python-dotenv==1.2.1
Pillow==12.0.0
